[#29] chore: rename package me.linkkle.app
