+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.inandout",
+    package = "me.linkkle.app",
 )
 
 android_resource(
     name = "res",
-    package = "com.inandout",
+    package = "me.linkkle.app",
     res = "src/main/res",
 )
