+++ android/app/src/main/java/me/linkkle/app/MainActivity.java
@@ -1,4 +1,4 @@
-package com.inandout;
+package me.linkkle.app;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -7,7 +7,8 @@ import com.facebook.react.ReactRootView;
 public class MainActivity extends ReactActivity {
 
   /**
-   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * Returns the name of the main component registered from JavaScript. This is
+   * used to schedule
    * rendering of the component.
    */
   @Override
@@ -16,8 +17,10 @@ public class MainActivity extends ReactActivity {
   }
 
   /**
-   * Returns the instance of the {@link ReactActivityDelegate}. There the RootView is created and
-   * you can specify the renderer you wish to use - the new renderer (Fabric) or the old renderer
+   * Returns the instance of the {@link ReactActivityDelegate}. There the RootView
+   * is created and
+   * you can specify the renderer you wish to use - the new renderer (Fabric) or
+   * the old renderer
    * (Paper).
    */
   @Override
@@ -40,7 +43,8 @@ public class MainActivity extends ReactActivity {
 
     @Override
     protected boolean isConcurrentRootEnabled() {
-      // If you opted-in for the New Architecture, we enable Concurrent Root (i.e. React 18).
+      // If you opted-in for the New Architecture, we enable Concurrent Root (i.e.
+      // React 18).
       // More on this on https://reactjs.org/blog/2022/03/29/react-v18.html
       return BuildConfig.IS_NEW_ARCHITECTURE_ENABLED;
     }
