+++ android/app/src/main/java/me/linkkle/app/newarchitecture/MainApplicationReactNativeHost.java
@@ -1,4 +1,4 @@
-package com.inandout.newarchitecture;
+package me.linkkle.app.newarchitecture;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -19,17 +19,20 @@ import com.facebook.react.fabric.CoreComponentsRegistry;
 import com.facebook.react.fabric.FabricJSIModuleProvider;
 import com.facebook.react.fabric.ReactNativeConfig;
 import com.facebook.react.uimanager.ViewManagerRegistry;
-import com.inandout.BuildConfig;
-import com.inandout.newarchitecture.components.MainComponentsRegistry;
-import com.inandout.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
+import me.linkkle.app.BuildConfig;
+import me.linkkle.app.newarchitecture.components.MainComponentsRegistry;
+import me.linkkle.app.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * A {@link ReactNativeHost} that helps you load everything needed for the New Architecture, both
+ * A {@link ReactNativeHost} that helps you load everything needed for the New
+ * Architecture, both
  * TurboModule delegates and the Fabric Renderer.
  *
- * <p>Please note that this class is used ONLY if you opt-in for the New Architecture (see the
+ * <p>
+ * Please note that this class is used ONLY if you opt-in for the New
+ * Architecture (see the
  * `newArchEnabled` property). Is ignored otherwise.
  */
 public class MainApplicationReactNativeHost extends ReactNativeHost {
@@ -45,11 +48,14 @@ public class MainApplicationReactNativeHost extends ReactNativeHost {
   @Override
   protected List<ReactPackage> getPackages() {
     List<ReactPackage> packages = new PackageList(this).getPackages();
-    // Packages that cannot be autolinked yet can be added manually here, for example:
-    //     packages.add(new MyReactNativePackage());
-    // TurboModules must also be loaded here providing a valid TurboReactPackage implementation:
-    //     packages.add(new TurboReactPackage() { ... });
-    // If you have custom Fabric Components, their ViewManagers should also be loaded here
+    // Packages that cannot be autolinked yet can be added manually here, for
+    // example:
+    // packages.add(new MyReactNativePackage());
+    // TurboModules must also be loaded here providing a valid TurboReactPackage
+    // implementation:
+    // packages.add(new TurboReactPackage() { ... });
+    // If you have custom Fabric Components, their ViewManagers should also be
+    // loaded here
     // inside a ReactPackage.
     return packages;
   }
@@ -61,9 +67,9 @@ public class MainApplicationReactNativeHost extends ReactNativeHost {
 
   @NonNull
   @Override
-  protected ReactPackageTurboModuleManagerDelegate.Builder
-      getReactPackageTurboModuleManagerDelegateBuilder() {
-    // Here we provide the ReactPackageTurboModuleManagerDelegate Builder. This is necessary
+  protected ReactPackageTurboModuleManagerDelegate.Builder getReactPackageTurboModuleManagerDelegateBuilder() {
+    // Here we provide the ReactPackageTurboModuleManagerDelegate Builder. This is
+    // necessary
     // for the new architecture and to use TurboModules correctly.
     return new MainApplicationTurboModuleManagerDelegate.Builder();
   }
@@ -98,9 +104,8 @@ public class MainApplicationReactNativeHost extends ReactNativeHost {
 
                 final ReactInstanceManager reactInstanceManager = getReactInstanceManager();
 
-                ViewManagerRegistry viewManagerRegistry =
-                    new ViewManagerRegistry(
-                        reactInstanceManager.getOrCreateViewManagers(reactApplicationContext));
+                ViewManagerRegistry viewManagerRegistry = new ViewManagerRegistry(
+                    reactInstanceManager.getOrCreateViewManagers(reactApplicationContext));
 
                 return new FabricJSIModuleProvider(
                     reactApplicationContext,
