+++ android/app/src/main/java/me/linkkle/app/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.inandout.newarchitecture.components;
+package me.linkkle.app.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
@@ -6,11 +6,15 @@ import com.facebook.react.fabric.ComponentFactory;
 import com.facebook.soloader.SoLoader;
 
 /**
- * Class responsible to load the custom Fabric Components. This class has native methods and needs a
- * corresponding C++ implementation/header file to work correctly (already placed inside the jni/
+ * Class responsible to load the custom Fabric Components. This class has native
+ * methods and needs a
+ * corresponding C++ implementation/header file to work correctly (already
+ * placed inside the jni/
  * folder for you).
  *
- * <p>Please note that this class is used ONLY if you opt-in for the New Architecture (see the
+ * <p>
+ * Please note that this class is used ONLY if you opt-in for the New
+ * Architecture (see the
  * `newArchEnabled` property). Is ignored otherwise.
  */
 @DoNotStrip
@@ -19,7 +23,8 @@ public class MainComponentsRegistry {
     SoLoader.loadLibrary("fabricjni");
   }
 
-  @DoNotStrip private final HybridData mHybridData;
+  @DoNotStrip
+  private final HybridData mHybridData;
 
   @DoNotStrip
   private native HybridData initHybrid(ComponentFactory componentFactory);
