+++ android/app/src/main/jni/MainComponentsRegistry.h
@@ -5,28 +5,31 @@
 #include <react/renderer/componentregistry/ComponentDescriptorProviderRegistry.h>
 #include <react/renderer/componentregistry/ComponentDescriptorRegistry.h>
 
-namespace facebook {
-namespace react {
+namespace facebook
+{
+    namespace react
+    {
 
-class MainComponentsRegistry
-    : public facebook::jni::HybridClass<MainComponentsRegistry> {
- public:
-  // Adapt it to the package you used for your Java class.
-  constexpr static auto kJavaDescriptor =
-      "Lcom/inandout/newarchitecture/components/MainComponentsRegistry;";
+        class MainComponentsRegistry
+            : public facebook::jni::HybridClass<MainComponentsRegistry>
+        {
+        public:
+            // Adapt it to the package you used for your Java class.
+            constexpr static auto kJavaDescriptor =
+                "Lme.linkkle.app/newarchitecture/components/MainComponentsRegistry;";
 
-  static void registerNatives();
+            static void registerNatives();
 
-  MainComponentsRegistry(ComponentFactory *delegate);
+            MainComponentsRegistry(ComponentFactory *delegate);
 
- private:
-  static std::shared_ptr<ComponentDescriptorProviderRegistry const>
-  sharedProviderRegistry();
+        private:
+            static std::shared_ptr<ComponentDescriptorProviderRegistry const>
+            sharedProviderRegistry();
 
-  static jni::local_ref<jhybriddata> initHybrid(
-      jni::alias_ref<jclass>,
-      ComponentFactory *delegate);
-};
+            static jni::local_ref<jhybriddata> initHybrid(
+                jni::alias_ref<jclass>,
+                ComponentFactory *delegate);
+        };
 
-} // namespace react
+    } // namespace react
 } // namespace facebook
